requirements.txt
     django
     djangorestframework
     django-extensions
     django-cors-headers
     django-rest-knox
     cryptography
     jinja2
     boto3
     pandas
     numpy
     scipy